fix(datafusion): drop now-redundant index row-count conversions
Widening `IndexFileMeta.row_count` to `i64` turns the datafusion-side
`i64::from(...)` and `as i64` on that field into no-ops, which
`clippy::useless_conversion` / `clippy::unnecessary_cast` reject under
`-D warnings`. The CI `check` job runs clippy over the whole workspace,
so these sites are gated even though the type change lives in `paimon`.

Only the lib site was reported: clippy stops at the first failing target,
so the two test sites would have surfaced on the following run.

No behaviour change -- every removed conversion was already identity.

+++ crates/integrations/datafusion/src/system_tables/table_indexes.rs
@@ -132,7 +132,7 @@ impl TableProvider for TableIndexesTable {
             index_types.push(index_file.index_type.as_str());
             file_names.push(index_file.file_name.as_str());
             file_sizes.push(index_file.file_size);
-            row_counts.push(i64::from(index_file.row_count));
+            row_counts.push(index_file.row_count);
             append_dv_ranges(
                 &mut dv_ranges,
                 index_file
